+++ Project Portfolio V4/BHB.Common/API/ApiResponse.cs
@@ -1,12 +1,5 @@
-﻿using System.Collections;
-
-namespace BHB.Common.API
+﻿namespace BHB.Common.API
 {
-    public enum Status
-    {
-        Success,
-        Failure
-    }
     public class ApiResponse
     {
         private string _message;
@@ -22,12 +15,5 @@ namespace BHB.Common.API
     public class ApiResponse<T> :ApiResponse
     {
         public T Data { get; set; }
-    }
-
-    public class PagedApiResponse<TList> : ApiResponse<TList> where TList : IEnumerable
-    {
-        public int Page { get; set; }
-        public int NumPerPage { get; set; }
-        public int TotalPages { get; set; }
     }
 }
