+++ Project Portfolio V4/BHB.Common/API/PagedApiResponse.cs
@@ -1,0 +1,11 @@
+﻿using System.Collections;
+
+namespace BHB.Common.API
+{
+    public class PagedApiResponse<TList> : ApiResponse<TList> where TList : IEnumerable
+    {
+        public int Page { get; set; }
+        public int NumPerPage { get; set; }
+        public int TotalPages { get; set; }
+    }
+}
