+++ Project Portfolio V4/BHB.Common/API/Status.cs
@@ -1,0 +1,8 @@
+﻿namespace BHB.Common.API
+{
+    public enum Status
+    {
+        Success,
+        Failure
+    }
+}
